requirements.txt
duckduckgo-search
requests
beautifulsoup4
textblob
apify-client
python-dotenv
